GH-586: Override fixedSizeBinary method for UnionMapWriter (#885)
## What's Changed

`UnionMapWriter` will null out the entire map struct entry instead of
setting the value to null in:
```
childWriter.value().fixedSizeBinary().writeNull();
```

This PR overrides the `fixedSizeBinary()` method for the
`UnionMapWriter`, resolving it.
In addition, it introduces the `fixedSizeBinary(int byteWidth)` which
needs to be used to initialize `key` and `value` writes in
`UnionMapWriter` if they have not been initialized.

Closes #586.